data_warehouse_api_automation/transform-load-weather-data/requirements.txt
# Function dependencies, for example:
# package>=version
  
google-cloud-storage==1.13.1
google-cloud-bigquery==1.8.1
google-cloud-core==0.29.1